backend/pyproject.toml
[project]
name = "photo-service"
version = "0.1.0"
requires-python = ">=3.12"
dependencies = [
  "fastapi>=0.111",
  "uvicorn[standard]>=0.29",
  "sqlalchemy[asyncio]>=2.0",
  "asyncpg>=0.29",
  "alembic>=1.13",
  "pydantic-settings>=2.2",
  "passlib[bcrypt]>=1.7",
  "python-jose[cryptography]>=3.3",
  "python-multipart>=0.0.9",
  "pillow>=10.3",
  "piexif>=1.1",
  "celery[redis]>=5.3",
  "redis>=5.0",
  "geopy>=2.4",
  "pyheif>=0.7",
]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "httpx>=0.27",
  "anyio>=4",
  "aiosqlite>=0.20",
]
all = [
  "photo-service[dev]",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.pytest.ini_options]
asyncio_mode = "auto"
